Improved README: grammar and organization.
The README is very complete and useful but had some grammatical errors
and was organized in such a way that the laundry list of options took
precedence over the use examples.  There was also a typo in the example
for exludeFiles which showed using an object when, in fact, it should be
an array.