Apply file attributes before post-install scripts.
When file attributes are specified in the options (mode, owner, and
group), these are translated into commands that run in the post-install
section (%post) of the spec file.  This change ensures that these
attribute changes are run before any user-specified post-install scripts
in the case that these scripts require certain permissions or user/group
membership to run correctly.