Removing bump package.
This is a bit of an opinionated change but it is nicer to prepare a
release with the version bumped in a separate branch then tag it after
being merged into master.  Otherwise we have a separate commit for the
bump after the merge occurred even though the release was prepared.