Split AutoReqProv into AutoReq and AutoProv tags.
The documentation shows that three tags for automatic processing exist:
AutoReqProv, AutoReq, and AutoProv.  AutoReqProv sets both AutoReq and
AutoProv with its value.  Thus, the spec now has two tags: autoReq and
autoProv.  When both of these are false, the AutoReqProv tag is used.
Otherwise, the individual tags are used if only one is false.